panic commit. charger broke

+++ src/main/webapp/jsp/productManagement/view.jsp
@@ -9,6 +9,7 @@
     String applicationMessage = (String) request.getAttribute("applicationMessage");
     String menuActiveLink = "Prodotti";
     Product[] products = (Product[]) request.getAttribute("products");
+    Boolean soldProductsAction = (Boolean) request.getAttribute("soldProductsAction");
 %>
 
 <!DOCTYPE html>
@@ -18,6 +19,12 @@
         <script>
             function insertProduct(){
                 document.insertForm.submit();
+            }
+            function viewProduct(){
+                document.viewProductForm.submit();
+            }
+            function viewSoldProducts(){
+                document.viewSoldProductsForm.submit();
             }
             function mainOnLoadHandler(){
                 document.querySelector("#newProductButton").addEventListener("click",insertProduct);
@@ -38,8 +45,13 @@
                 transition: background-color 0.3s ease;
                 background-color: #28a745;
             }
+            #buttons{
+                display: flex;
+                flex-direction: row;
+                margin-bottom: 12px;
+            }
             #insertProductButtonSelection{
-                margin: 12px 0;
+                margin-right: 12px;
             }
             #products{
                 margin: 12px 0;
@@ -72,13 +84,35 @@
     <body>
         <%@include file="/include/header.jsp"%>
     <main>
+        <%if(!soldProductsAction){%>
+            <section id="pageTitle">
+                <h1>Centro Gestione dei Prodotti</h1>
+            </section>
+        <%}%>
+        <%if(soldProductsAction){%>
         <section id="pageTitle">
-            <h1>Centro Gestione dei Prodotti</h1>
+            <h1>Elenco dei Prodotti Venduti</h1>
         </section>
+        <%}%>
 
-        <section id="insertProductButtonSelection">
-            <input type="button" id="insertProductButton" name="insertProductButton"
-                   class="button" value="Inserisci un Prodotto nel tuo Account" onclick="insertProduct()"/>
+        <section id="buttons">
+            <article id="insertProductButtonSelection">
+                <input type="button" id="insertProductButton" name="insertProductButton"
+                       class="button" value="Inserisci un Prodotto nel tuo Account" onclick="insertProduct()"/>
+            </article>
+
+            <%if(!soldProductsAction){%>
+                <article id="viewSoldProductsSelection">
+                    <input type="button" id="viewSoldProductsButton" name="viewSoldProductsButton"
+                           class="button" value="Visualizza i Prodotti Venduti" onclick="viewSoldProducts()"/>
+                </article>
+            <%}%>
+            <%if(soldProductsAction){%>
+                <article id="viewProductsSelection">
+                    <input type="button" id="viewProductsButton" name="viewProductsButton"
+                           class="button" value="Visualizza i prodotti nel tuo account" onclick="viewProduct()"/>
+                </article>
+            <%}%>
         </section>
 
         <section id="productListBreak">
@@ -89,12 +123,19 @@
             <section id="products" class="clearfix">
                 <%for (i = 0; i < products.length; i++){%>
                 <article>
-                    <a href="javascript:deleteProduct(<%=products[i].getProductID()%>)">
-                        <img id="trashcan" src="images/trashcan.png" width="24" height="24" alt="X">
-                    </a>
+                    <%if(!soldProductsAction){%>
+                        <a href="javascript:deleteProduct(<%=products[i].getProductID()%>)">
+                            <img id="trashcan" src="images/trashcan.png" width="24" height="24" alt="X">
+                        </a>
+                    <%}%>
                     <b><span class="description"><%=products[i].getDescription()%></span></b>
                     <br/>
-                    <span class="current_price">Attualmente costa: &euro;<%=products[i].getCurrent_price()%></span>
+                    <%if(soldProductsAction){%>
+                        <span class="current_price">Venduto a: &euro;<%=products[i].getCurrent_price()%></span>
+                    <%}%>
+                    <%if(!soldProductsAction){%>
+                        <span class="current_price">Attualmente costa: &euro;<%=products[i].getCurrent_price()%></span>
+                    <%}%>
                 </article>
                 <%}%>
             </section>
@@ -102,6 +143,14 @@
 
         <form name="insertForm" method="post" action="Dispatcher">
             <input type="hidden" name="controllerAction" value="ProductManagement.insertView"/>
+        </form>
+
+        <form name="viewSoldProductsForm" method="post" action="Dispatcher">
+            <input type="hidden" name="controllerAction" value="ProductManagement.viewSoldProducts"/>
+        </form>
+
+        <form name="viewProductForm" method="post" action="Dispatcher">
+            <input type="hidden" name="controllerAction" value="ProductManagement.view"/>
         </form>
 
         <form name="deleteForm" method="post" action="Dispatcher">
